split functionality
