+++ std/jwt/signature.ts
@@ -1,0 +1,32 @@
+import type { Algorithm } from "./algorithm.ts"
+import { convertHexToBase64url, encrypt } from "./_util.ts";
+
+export async function create(
+  alg: Algorithm,
+  key: string,
+  input: string,
+): Promise<string> {
+  return convertHexToBase64url(await encrypt(alg, key, input));
+}
+
+export async function verify({
+  signature,
+  key,
+  alg,
+  signingInput,
+}: {
+  signature: string;
+  key: string;
+  alg: Algorithm;
+  signingInput: string;
+}): Promise<boolean> {
+  switch (alg) {
+    case "none":
+    case "HS256":
+    case "HS512": {
+      return signature === (await encrypt(alg, key, signingInput));
+    }
+    default:
+      throw new RangeError(`no matching crypto alg in the header: ${alg}`);
+  }
+}
